requirements.txt
Django==2.2.11
djangorestframework
gunicorn
psycopg2-binary